@@ -1,6 +1,7 @@
 package com.skilles.spokenword.mixin;
 
 import com.skilles.spokenword.util.ConfigUtil;
+import com.skilles.spokenword.util.Util;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayNetworkHandler;
@@ -9,8 +10,12 @@ import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.network.packet.s2c.play.CombatEventS2CPacket;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.listener.ClientPlayPacketListener;
+import net.minecraft.network.packet.s2c.play.DeathMessageS2CPacket;
+import net.minecraft.network.packet.s2c.play.EndCombatS2CPacket;
 import net.minecraft.network.packet.s2c.play.GameJoinS2CPacket;
+import net.minecraft.text.TranslatableText;
 import net.minecraft.util.math.BlockPos;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -40,10 +45,12 @@ public class NetworkHandlerMixin {
             sendMessages(ON_JOIN_LIST);
         }
     }
-    @Inject(method = "onCombatEvent", at = @At(value = "TAIL"))
-    private void onCombatInject(CombatEventS2CPacket packet, CallbackInfo ci) {
-        if(globalEnabled() && modeConfig().death && packet.type == CombatEventS2CPacket.Type.ENTITY_DIED) {
-            Entity attacker = world.getEntityById(packet.attackerEntityId);
+    @Inject(method = "onDeathMessage(Lnet/minecraft/network/packet/s2c/play/DeathMessageS2CPacket;)V", at = @At(value = "TAIL"))
+    private void onDeathMessage(DeathMessageS2CPacket packet, CallbackInfo ci) {
+        if(globalEnabled() && modeConfig().death) {
+            assert MinecraftClient.getInstance().world != null;
+            ClientWorld world = MinecraftClient.getInstance().world;
+            Entity attacker = world.getEntityById(packet.getKillerId());
             if(attacker != null) {
                 if(attacker instanceof MobEntity && deathConfig().pve) {
                     if(deathConfig().entityKillList.contains("ANY") || containsEntity(attacker, ListModes.HOSTILE)) {
@@ -52,10 +59,10 @@ public class NetworkHandlerMixin {
                         log("Add mob: " + attacker.getType().getName().getString());
                     }
                 } else if(attacker instanceof PlayerEntity && deathConfig().pvp) {
-                    sendMessages(attacker.getDisplayName().getString(), PVP_LIST);
+                    sendMessages(new TranslatableText("selectWorld.world"), PVP_LIST);
                 }
             } else {
-                sendMessages(world.getEntityById(packet.entityId), PVE_LIST);
+                sendMessages("The World", PVP_LIST);
             }
         }
     }
